chore: extract key human readable number function for reusability
