+++ webui/react/src/components/HumanReadableNumber.tsx
@@ -2,6 +2,7 @@ import { Tooltip } from 'antd';
 import React from 'react';
 
 import { CommonProps } from 'types';
+import { humanReadableNumber } from 'utils/number';
 
 interface Props extends CommonProps {
   num?: number | null;
@@ -17,20 +18,7 @@ const HumanReadableNumber: React.FC<Props> = ({
   if (num == null) return null;
 
   const stringNum = num.toString();
-  let content: string = stringNum;
-
-  if (isNaN(num)) {
-    content = 'NaN';
-  } else if (!Number.isFinite(num)) {
-    content = `${num < 0 ? '-' : ''}Infinity`;
-  } else if (!Number.isInteger(num)) {
-    content = num.toFixed(precision);
-
-    const absoluteNum = Math.abs(num);
-    if (absoluteNum < 0.01 || absoluteNum > 999) {
-      content = num.toExponential(precision);
-    }
-  }
+  const content = humanReadableNumber(num, precision);
 
   return (
     <Tooltip title={`${tooltipPrefix}${stringNum}`}>
